+++ API/.env
@@ -6,6 +6,7 @@ DB_PORT_CLOUD=6265
 
 HOST_FRONT_URL=http://localhost:5173
 AUTH0_CLIENT_ID=0VbI0VgjMwnZ0YhQZZ4tJSjfcc7v7jaq
+SECRET_TOKEN_AUTH0=-OVdWrqmPDOF-0SPFpwybF1QTfe3mUOSuHW3Xqn2SitSPLRK0zpLI_1K-Wn41mrL
 
 TOKEN_ACCESS_MERCADO_PAGO=TEST-2452790176162293-070201-1acb740989b97e709a2c5e442aa179d2-232893913
